% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msaenet-package.R
\docType{package}
\name{msaenet-package}
\alias{msaenet-package}
\title{msaenet: Multi-Step Adaptive Estimation Methods for Sparse Regressions}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Multi-step adaptive elastic-net (MSAENet) algorithm for feature selection in high-dimensional regressions proposed in Xiao and Xu (2015) \doi{10.1080/00949655.2015.1016944}, with support for multi-step adaptive MCP-net (MSAMNet) and multi-step adaptive SCAD-net (MSASNet) methods.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://nanx.me/msaenet/}
  \item \url{https://github.com/nanxstats/msaenet}
  \item Report bugs at \url{https://github.com/nanxstats/msaenet/issues}
}

}
\author{
\strong{Maintainer}: Nan Xiao \email{me@nanx.me} (\href{https://orcid.org/0000-0002-0250-5673}{ORCID})

Authors:
\itemize{
  \item Qing-Song Xu \email{qsxu@csu.edu.cn}
}

}
\keyword{internal}
