% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msaenet-eval.R
\name{msaenet.mae}
\alias{msaenet.mae}
\title{Mean Absolute Error (MAE)}
\usage{
msaenet.mae(yreal, ypred)
}
\arguments{
\item{yreal}{Vector. True response.}

\item{ypred}{Vector. Predicted response.}
}
\value{
MAE
}
\description{
Compute mean absolute error (MAE).
}
\author{
Nan Xiao <\url{http://nanx.me}>
}

