% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msaenet-eval.R
\name{msaenet.mse}
\alias{msaenet.mse}
\title{Mean Squared Error (MSE)}
\usage{
msaenet.mse(yreal, ypred)
}
\arguments{
\item{yreal}{Vector. True response.}

\item{ypred}{Vector. Predicted response.}
}
\value{
MSE
}
\description{
Compute mean squared error (MSE).
}
\author{
Nan Xiao <\url{https://nanx.me}>
}
