% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msaenet.eval.R
\name{msaenet.rmse}
\alias{msaenet.rmse}
\title{Root Mean Squared Error (RMSE)}
\usage{
msaenet.rmse(yreal, ypred)
}
\arguments{
\item{yreal}{Vector. True response.}

\item{ypred}{Vector. Predicted response.}
}
\value{
RMSE
}
\description{
Compute root mean squared error (RMSE).
}
\author{
Nan Xiao <\url{http://nanx.me}>
}

