% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msaenet-eval.R
\name{msaenet.rmsle}
\alias{msaenet.rmsle}
\title{Root Mean Squared Logarithmic Error (RMSLE)}
\usage{
msaenet.rmsle(yreal, ypred)
}
\arguments{
\item{yreal}{Vector. True response.}

\item{ypred}{Vector. Predicted response.}
}
\value{
RMSLE
}
\description{
Compute root mean squared logarithmic error (RMSLE).
}
\author{
Nan Xiao <\url{http://nanx.me}>
}

