% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msaenet-tune.R
\name{msaenet.tune.glmnet.alpha}
\alias{msaenet.tune.glmnet.alpha}
\title{Automatic alpha tuning by k-fold cross-validation}
\usage{
msaenet.tune.glmnet.alpha(..., alphas, seed, parallel)
}
\value{
best model object and best alpha value
}
\description{
Automatic alpha tuning by k-fold cross-validation
}
\author{
Nan Xiao <\url{http://nanx.me}>
}
\keyword{internal}

