% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msaenet-tune-glmnet.R
\name{msaenet.tune.glmnet}
\alias{msaenet.tune.glmnet}
\title{Automatic parameter tuning for glmnet models}
\usage{
msaenet.tune.glmnet(x, y, family, alphas, nfolds, rule, seed, parallel, ...)
}
\value{
best model object and best alpha value
}
\description{
Automatic parameter tuning for glmnet models
}
\author{
Nan Xiao <\url{https://nanx.me}>
}
\keyword{internal}

