% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msaenet-tune-ncvreg.R
\name{msaenet.tune.ncvreg}
\alias{msaenet.tune.ncvreg}
\title{Automatic (parallel) parameter tuning for ncvreg models}
\usage{
msaenet.tune.ncvreg(
  x,
  y,
  family,
  penalty,
  gammas,
  alphas,
  tune,
  nfolds,
  ebic.gamma,
  eps,
  max.iter,
  seed,
  parallel,
  ...
)
}
\value{
Optimal model object, parameter set, and criterion value
}
\description{
Automatic (parallel) parameter tuning for ncvreg models
}
\references{
Chen, Jiahua, and Zehua Chen. (2008).
Extended Bayesian information criteria for model selection with
large model spaces. \emph{Biometrika} 95(3), 759--771.
}
\author{
Nan Xiao <\url{https://nanx.me}>
}
\keyword{internal}
