% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msaenet-tune-ncvreg.R
\name{msaenet.tune.nsteps.ncvreg}
\alias{msaenet.tune.nsteps.ncvreg}
\title{Select the number of adaptive estimation steps}
\usage{
msaenet.tune.nsteps.ncvreg(model.list, tune.nsteps, ebic.gamma.nsteps)
}
\value{
optimal step number
}
\description{
Select the number of adaptive estimation steps
}
\author{
Nan Xiao <\url{https://nanx.me}>
}
\keyword{internal}
