% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msamnet.R
\name{msamnet}
\alias{msamnet}
\title{Multi-Step Adaptive MCP-Net}
\usage{
msamnet(x, y, family = c("gaussian", "binomial", "poisson", "cox"),
  init = c("mnet", "ridge"), nsteps = 2L, nfolds = 5L, gammas = c(1.01,
  1.7, 3, 100), alphas = seq(0.05, 0.95, 0.05), gamma = 1, seed = 1001,
  parallel = FALSE, verbose = FALSE)
}
\arguments{
\item{x}{Data matrix.}

\item{y}{Response vector if \code{family} is \code{"gaussian"},
\code{"binomial"} or \code{"poisson"}.
If \code{family} is \code{"cox"}, a response matrix made by
\code{\link[survival]{Surv}}.}

\item{family}{Model family, can be \code{"gaussian"},
\code{"binomial"}, \code{"poisson"}, or \code{"cox"}.}

\item{init}{Type of the penalty used in the initial
estimation step. Can be \code{"mnet"} or \code{"ridge"}.}

\item{nsteps}{How many adaptive estimation steps? At least 2.
We assume adaptive MCP-net has only 1 adaptive estimation step.}

\item{nfolds}{Fold numbers of cross-validation.}

\item{gammas}{Vector of candidate \code{gamma}s to use in MCP-Net.}

\item{alphas}{Vector of candidate \code{alpha}s to use in MCP-Net.}

\item{gamma}{Scaling factor for adaptive weights:
\code{weights = coefs^(-gamma)}.}

\item{seed}{Two random seeds for cross-validation fold division
in two estimation steps.}

\item{parallel}{Logical. Enable parallel parameter tuning or not,
default is {FALSE}. To enable parallel tuning, load the
\code{doParallel} package and run \code{registerDoParallel()}
with the number of CPU cores before calling this function.}

\item{verbose}{Should we print out the estimation progress?}
}
\value{
List of coefficients \code{beta} and
\code{ncvreg} model object \code{model}.
}
\description{
Multi-Step Adaptive MCP-Net
}
\examples{
dat = msaenet.sim.gaussian(n = 150, p = 500, rho = 0.6,
                           coef = rep(1, 5), snr = 2, p.train = 0.7,
                           seed = 1001)

msamnet.fit = msamnet(dat$x.tr, dat$y.tr,
                      gammas = 3, alphas = seq(0.3, 0.9, 0.3),
                      nsteps = 3L, seed = 1003)

print(msamnet.fit)
msaenet.nzv(msamnet.fit)
msaenet.fp(msamnet.fit, 1:5)
msaenet.tp(msamnet.fit, 1:5)
msamnet.pred = predict(msamnet.fit, dat$x.te)
msaenet.rmse(dat$y.te, msamnet.pred)
plot(msamnet.fit)
}
\author{
Nan Xiao <\url{http://nanx.me}>
}

