% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-msaenet.R
\name{print.msaenet}
\alias{print.msaenet}
\title{Print AENet/MSAENet Model Information}
\usage{
\method{print}{msaenet}(x, ...)
}
\arguments{
\item{x}{An object of class \code{msaenet}.}

\item{...}{Additional parameters for \code{\link{print}}.}
}
\description{
Print AENet/MSAENet model objects
(currently, only printing the model from the final step).
}
\examples{
dat = msaenet.sim.gaussian(n = 150, p = 500, rho = 0.6,
                           coef = rep(1, 5), snr = 2, p.train = 0.7,
                           seed = 1001)

msaenet.fit = msaenet(dat$x.tr, dat$y.tr,
                      alphas = seq(0.2, 0.8, 0.2),
                      nsteps = 3L, seed = 1003)

print(msaenet.fit)
}
\author{
Nan Xiao <\url{http://nanx.me}>
}

