% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msaenet-package.R
\docType{package}
\name{msaenet-package}
\alias{msaenet-package}
\title{Multi-Step Adaptive Elastic-Net}
\description{
Multi-Step Adaptive Elastic-Net
}
\details{
The vignette can be opened with \code{vignette("msaenet")}.

\tabular{ll}{ Package: \tab msaenet\cr
Type: \tab Package\cr
License: \tab GPL\cr }
}
\author{
Nan Xiao <\email{me@nanx.me}>
}

