% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msaenet-package.R
\docType{package}
\name{msaenet-package}
\alias{msaenet-package}
\title{Multi-Step Adaptive Estimation Methods for Sparse Regressions}
\description{
Multi-step adaptive elastic-net (MSAENet) algorithm for
feature selection in high-dimensional regressions proposed in
Xiao and Xu (2015) <DOI:10.1080/00949655.2015.1016944>,
with support for multi-step adaptive MCP-Net (MSAMNet) and
multi-step adaptive SCAD-Net (MSASNet) methods.
}
\details{
Browse the vignette with \code{vignette("msaenet")}.

\tabular{ll}{
Package: \tab msaenet\cr
Type: \tab Package\cr
License: \tab GPL-3}
}
\references{
Nan Xiao and Qing-Song Xu. (2015). Multi-step adaptive elastic-net:
reducing false positives in high-dimensional variable selection.
\emph{Journal of Statistical Computation and Simulation} 85(18), 3755--3765.
}
\author{
Nan Xiao <\email{me@nanx.me}>
}
