% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msaenet-tune-glmnet.R
\name{msaenet.tune.glmnet}
\alias{msaenet.tune.glmnet}
\title{Automatic parameter tuning for glmnet by k-fold cross-validation}
\usage{
msaenet.tune.glmnet(..., alphas, seed, parallel)
}
\value{
best model object and best alpha value
}
\description{
Automatic parameter tuning for glmnet by k-fold cross-validation
}
\author{
Nan Xiao <\url{http://nanx.me}>
}
\keyword{internal}

