% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msaenet-tune-glmnet.R
\name{msaenet.tune.glmnet}
\alias{msaenet.tune.glmnet}
\title{Automatic (parallel) parameter tuning for glmnet models}
\usage{
msaenet.tune.glmnet(x, y, family, alphas, tune, nfolds, rule, ebic.gamma, seed,
  parallel, ...)
}
\value{
Optimal model object, parameter set, and criterion value
}
\description{
Automatic (parallel) parameter tuning for glmnet models
}
\references{
Chen, Jiahua, and Zehua Chen. (2008).
Extended Bayesian information criteria for model selection with
large model spaces. \emph{Biometrika} 95(3), 759--771.
}
\author{
Nan Xiao <\url{https://nanx.me}>
}
\keyword{internal}
