% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msaenet-tune-ncvreg.R
\name{msaenet.tune.ncvreg}
\alias{msaenet.tune.ncvreg}
\title{Automatic parameter tuning for ncvreg by k-fold cross-validation}
\usage{
msaenet.tune.ncvreg(..., family, gammas, alphas, seed, parallel)
}
\value{
best model object, best gamma, and best alpha
}
\description{
Automatic parameter tuning for ncvreg by k-fold cross-validation
}
\author{
Nan Xiao <\url{http://nanx.me}>
}
\keyword{internal}

